チームリストをコンポーネント化した
